+++ allsecureexchange/upload/catalog/language/sr-RS/extension/payment/allsecure_exchange.php
@@ -1,0 +1,72 @@
+<?php
+$_['order_cancelled'] = 'Porudžbina je otkazana';
+$_['order_error'] = 'Transakcija nije bila uspešna. Molimo pokušajte ponovo.';
+$_['order_on_hold'] = 'Čeka se zaključenje ili otkazivanje.';
+$_['button_confirm'] = 'Potvrda porudžbine';
+
+$_['redirect_text'] = 'Preusmeravanje. Molimo sačekajte.';
+$_['loading_text'] = 'Forma za plaćanje se učitava.';
+
+$_['card_holder'] = 'Korisnik kartice';
+$_['card_number'] = 'Broj kartice';
+$_['card_ccv'] = 'CVV';
+$_['card_expiry_month'] = 'Mesec';
+$_['card_expiry_year'] = 'Godina';
+
+/* Payment Success Page */
+$_['payment_success_allsecure_exchange_heading_title'] = 'Vaša porudžbina je uspešna!';
+$_['transaction_details'] = 'Detalji transakcije';
+$_['transaction_type'] = 'Tip transakcije';
+$_['transaction_payment_method'] = 'Metod plaćanja';
+$_['transaction_amount'] = 'Iznos';
+$_['transaction_authcode'] = 'Kod';
+
+/* Card decline messages */
+$_['Error1000'] = 'CONFIG ERROR. Some fundamental error in your request';
+$_['Error1001'] = 'CONFIG ERROR. The upstream system responded with an unknown response';
+$_['Error1002'] = 'CONFIG ERROR. Request data are malformed or missing';
+$_['Error1003'] = 'CONFIG ERROR. Transaction could not be processed';
+$_['Error1004'] = 'CONFIG ERROR. The request signature you provided was wrong';
+$_['Error1005'] = 'CONFIG ERROR. The XML you provided was malformed or invalid';
+$_['Error1006'] = 'CONFIG ERROR. Preconditions failed, e.g. capture on a failed authorize';
+$_['Error1007'] = 'CONFIG ERROR. Something is wrong your configuration, please contact your integration engineer';
+$_['Error1008'] = 'CONFIG ERROR. Unexpected system error';
+$_['Error9999'] = 'CONFIG ERROR. We received an error which is not (yet) mapped to a better error code';
+$_['Error2001'] = 'Kartica koju ste uneli ne može biti prihvacena. Molimo pokušajte drugom karticom';
+$_['Error2002'] = 'Transakcija je otkazana od strane kupca';
+$_['Error2003'] = 'Vaša transakcija nije odobrena. Pokušajte ponovo ili promenite karticu';
+$_['Error2004'] = 'Vaša transakcija nije odobrena. Pokušajte ponovo ili promenite karticu';
+$_['Error2005'] = 'Korisnik nije u predviđenom roku završio transakciju';
+$_['Error2006'] = 'Vaša transakcija nije odobrena. Pokušajte ponovo ili promenite karticu';
+$_['Error2007'] = 'Uneti podaci o platnoj kartici nisu tacni. Proverite podatke i pokušajte ponovo';
+$_['Error2008'] = 'Kartica koju ste uneli ne može biti prihvacena. Molimo pokušajte drugom karticom';
+$_['Error2009'] = 'Kartica koju ste uneli je istekla. Molimo izaberite drugu karticu';
+$_['Error2010'] = 'Vaša transakcija nije odobrena. Molimo proverite sa bankom koja vam je izdala karticu';
+$_['Error2011'] = 'Kartica koju ste uneli ne može biti prihvacena. Molimo pokušajte drugom karticom';
+$_['Error2012'] = 'Transakcija je otkazana';
+$_['Error2013'] = 'Vaša transakcija nije odobrena. Molimo izaberite drugu karticu ili proverite sa bankom koja vam je izdala karticu';
+$_['Error2014'] = 'Vaša transakcija nije odobrena. Molimo izaberite drugu karticu ili proverite sa bankom koja vam je izdala karticu';
+$_['Error2015'] = 'Vaša transakcija nije odobrena. Molimo izaberite drugu karticu ili proverite sa bankom koja vam je izdala karticu';
+$_['Error2016'] = 'Vaša transakcija nije odobrena. Molimo izaberite drugu karticu ili proverite sa bankom koja vam je izdala karticu';
+$_['Error2017'] = 'IBAN nije validan. Proverite podatke i pokušajte ponovo';
+$_['Error2018'] = 'BIC nije validan. Proverite podatke i pokušajte ponovo';
+$_['Error2019'] = 'Podaci o korisniku nisu validni. Proverite podatke i pokušajte ponovo';
+$_['Error2020'] = 'CVV kod je obavezan. Proverite podatke i pokušajte ponovo';
+$_['Error2021'] = 'Vaša 3D Secure autentifikacija nije uspela. Molimo pozovite banku koja vam je izdala karticu ili koristite karticu koja nije 3D-Secure';
+$_['Error3001'] = 'KOMUNIKACIJA: Transakcija je istekla. Napravite pauzu i pokušajte nešto kasnije';
+$_['Error3002'] = 'KOMUNIKACIJA: Transakcija nije dozvoljena';
+$_['Error3003'] = 'KOMUNIKACIJA: Sistem privremeno nedostupan. Napravite pauzu i pokušajte nešto kasnije';
+$_['Error3004'] = 'ID Transakcije već korišćen';
+$_['Error3005'] = 'KOMUNIKACIJA: Problem u konekciji sa bankom. Napravite pauzu i pokušajte nešto kasnije';
+$_['Error7001'] = 'Zahtev za planiranjem transakcije nije validan';
+$_['Error7002'] = 'Zahtev za planiranjem transakcije nije prihvaćen';
+$_['Error7005'] = 'Planirana transakcija nije validna';
+$_['Error7010'] = 'RegistrationId je neophodan';
+$_['Error7020'] = 'RegistrationId nije validan';
+$_['Error7030'] = 'RegistrationId nema odgovarajuću referencu';
+$_['Error7035'] = 'Inicijalna transakcija kojom započinju planirana ponavljajuća plaćanja mora biti "register", "debit+register" ili "preuth+register"';
+$_['Error7036'] = 'Period između inicijalne i sledeće transakcije mora biti duži od 24 sata';
+$_['Error7040'] = 'ScheduleId (id planirane transakcije) nije validan ili nije povezan na ovaj kanal plaćanja';
+$_['Error7050'] = 'Planirana početna transakcija (startDateTime) nije validnog formata ili je starija od 24h';
+$_['Error7060'] = 'Planirana naredna transakcija (continueDateTime ) nije validnog formata ili je starija od 24h';
+$_['Error7070'] = 'Status planirane transakcije nije validnog formata za zahtevanu operaciju';