+++ allsecureexchange/upload/system/library/allsecure-exchange/plugin/plugin.php
@@ -6,10 +6,12 @@ final class AllsecureExchangePlugin
 {
 	const METHOD_DEBIT = 'debit';
     const METHOD_PREAUTHORIZE = 'preauthorize';
+	const HOST_TEST = 'https://asxgw.paymentsandbox.cloud/';
+    const HOST_LIVE = 'https://asxgw.com/';
 
     public function getVersion()
     {
-        return '1.4.0';
+        return '1.5.0';
     }
 
     public function getName()